+++ app/Game/Werewolf/Command/Witch.hs
@@ -23,6 +23,7 @@ import Control.Monad.Except
 import Control.Monad.Extra
 import Control.Monad.State
 
+import Data.Map  as Map
 import Data.Text (Text)
 
 import Game.Werewolf
@@ -33,11 +34,11 @@ import Game.Werewolf.Util
 healCommand :: Text -> Command
 healCommand callerName = Command $ do
     validateCommand callerName
-    whenM (use healUsed)                                    $ throwError [playerHasAlreadyHealedMessage callerName]
-    whenM (hasn'tuse $ events . traverse . _DevourEvent)    $ throwError [playerCannotDoThatRightNowMessage callerName]
+    whenM (use healUsed)        $ throwError [playerHasAlreadyHealedMessage callerName]
+    whenM (hasn'tuse votee)     $ throwError [playerCannotDoThatRightNowMessage callerName]
 
-    heal        .= True
     healUsed    .= True
+    votes       .= Map.empty
 
 passCommand :: Text -> Command
 passCommand callerName = Command $ do
@@ -48,9 +49,9 @@ passCommand callerName = Command $ do
 poisonCommand :: Text -> Text -> Command
 poisonCommand callerName targetName = Command $ do
     validateCommand callerName
-    whenM (use poisonUsed)                                              $ throwError [playerHasAlreadyPoisonedMessage callerName]
+    whenM (use poisonUsed)                      $ throwError [playerHasAlreadyPoisonedMessage callerName]
     validatePlayer callerName targetName
-    whenM (hasuse $ events . traverse . _DevourEvent . only targetName) $ throwError [playerCannotDoThatMessage callerName]
+    whenM (hasuse $ votee . named targetName)   $ throwError [playerCannotDoThatMessage callerName]
 
     poison      .= Just targetName
     poisonUsed  .= True